move __to_dict to NWBContainer
