+++ src/pynwb/file.py
@@ -11,7 +11,7 @@ from .epoch import Epoch
 from .ecephys import ElectrodeTable, ElectrodeTableRegion, ElectrodeGroup, Device
 from .icephys import IntracellularElectrode
 from .ophys import ImagingPlane
-from .core import NWBContainer
+from .core import NWBContainer, LabelledDict
 
 from h5py import RegionReference
 
@@ -30,22 +30,6 @@ class Image(Container):
 class SpecFile(Container):
     # TODO: Implement this
     pass
-
-
-class TimeSeriesDict(dict):
-    '''
-    A dict wrapper class for aggregating Timeseries
-    from the standard locations
-    '''
-
-    @docval({'name': 'ts_type', 'type': str, 'doc': 'the TimeSeries type ('})
-    def __init__(self, **kwargs):
-        ts_type = getargs('ts_type', kwargs)
-        self.__ts_type = ts_type
-
-    @property
-    def ts_type(self):
-        return self.__ts_type
 
 
 @register_class('NWBFile', CORE_NAMESPACE)
@@ -145,15 +129,15 @@ class NWBFile(NWBContainer):
         self.__stimulus = self.__build_ts('stimulus', kwargs)
         self.__stimulus_template = self.__build_ts('stimulus_template', kwargs)
 
-        self.__modules = self.__to_dict(getargs('modules', kwargs))
-        self.__epochs = self.__to_dict(getargs('epochs', kwargs))
+        self.__modules = self._to_dict('modules', kwargs)
+        self.__epochs = self._to_dict('epochs', kwargs)
         self.__ec_electrodes = getargs('ec_electrodes', kwargs)
-        self.__ec_electrode_groups = self.__to_dict(getargs('ec_electrode_groups', kwargs))
-        self.__devices = self.__to_dict(getargs('devices', kwargs))
+        self.__ec_electrode_groups = self._to_dict('ec_electrode_groups', kwargs)
+        self.__devices = self._to_dict('devices', kwargs)
 
-        self.__ic_electrodes = self.__to_dict(getargs('ic_electrodes', kwargs))
+        self.__ic_electrodes = self._to_dict('ic_electrodes', kwargs)
 
-        self.__imaging_planes = self.__to_dict(getargs('imaging_planes', kwargs))
+        self.__imaging_planes = self._to_dict('imaging_planes', kwargs)
 
         recommended = [
             'experimenter',
@@ -165,14 +149,11 @@ class NWBFile(NWBContainer):
         for attr in recommended:
             setattr(self, attr, kwargs.get(attr, None))
 
-    def __to_dict(self, arg):
-        if arg is None:
-            return dict()
-        else:
-            return {i.name: i for i in arg}
+    def _to_dict(self, label, const_args):
+        return super(NWBFile, self)._to_dict(getargs(label, const_args), label)
 
     def __build_ts(self, ts_type, kwargs):
-        ret = TimeSeriesDict(ts_type)
+        ret = LabelledDict(ts_type)
         const_arg = getargs(ts_type, kwargs)
         if const_arg:
             for ts in const_arg:
@@ -427,7 +408,7 @@ class NWBFile(NWBContainer):
 
     def __set_timeseries(self, ts_dict, ts, epoch=None):
         if ts.name in ts_dict:
-            msg = "%s already exists in %s" % (ts.name, ts_dict.ts_type)
+            msg = "%s already exists in %s" % (ts.name, ts_dict.label)
             raise ValueError(msg)
         ts_dict[ts.name] = ts
         if ts.parent is None:
